Update docs
Reorder to converter_reorder_attr
